Removed Spring Security, created rest of Controller and Service classes, got db tables created successfully
Removed Spring Security as it was causing issues with Postman requests saying Unauthorized. After removal and changing some join table names the tables were able to be created successfully. Created rest of the Service classes and finished working on ActivityService and ActivityController. Have to test each method next, and ActivityListController and ActivityListService are also done and need to be tested.