Added padding to td in all tables. Fixed update/add activity with Sets. Fixed some of the other code in ActivityController that was missing model data, added MuscleGroups to workout views
Add/Update Activity was not saving Sets. Was because the SetService was being used and it was returning int rather than Integer. Muscle Groups with count are shown in list_of_workouts and single_workout now as a quick way to gauge what workouts user has done. Included new query for it as well in ActivityRepository. Implemented new query in the ActivityTest as well.