Removed some files, for delete methods I redirected to the Get method rather than returning view.
Returning view was causing issues as some model data was not being passed. So I just set it up to redirect to the GET method instead. Have to be careful with this method because if you have a DELETE method with the same mapping it'll go to that one first based on the request code.